Extract notifications popover into NotificationsPopoverView.swift
UpdateTitlebarAccessory.swift was 3 lines over its length budget after
mounting the agent-count badges. Instead of ratcheting the budget up,
move the popover cluster (metrics, popover view, row, resize gripper,
hover tracking) into its own file and refresh the budget, shrinking the
god file from 3061 to 2465 lines.
